Initial commit to GitHub.
Added wide character support, please enable by defining "FINI_WIDE_SUPPORT".

Enables the use of C++ 0x features when available, improving performance.

Added handling for an additional hash comment as well as lines pre-ceeded with white space. Keys can therefore be indented.

Added a new get function, which will select a section in one procedure for convenience. The advantage this holds over the bracket operators is that it can return a default value if the section or key is not found.

Added the saving and parse as binary, this is useful for improving parsing speed, while decreasing file size and making it less convenient for a user to edit.

Conversion performance improvements for certain type overloading.

Added an example with a compilable Codeblocks project for GCC/MinGW/VC++, as well as a readme.